test: DiaryStorageManager 변경에 따른 DiaryStorageManagerTests 테스트 코드 수정
1. PersistentManagerTests -> DiaryStorageManagerTests 네이밍 변경
2. DiaryStorageManager 메서드 변경에 따른 DiaryStorageManagerTests 테스트 코드 변경
